added migration for seeding db data

+++ crud.py
@@ -1,3 +1,4 @@
+from datetime import datetime
 from sqlalchemy.orm import Session
 from fastapi import HTTPException
 
@@ -19,7 +20,8 @@ def get_events(db: Session):
 
 def make_reservation(reservation: schemas.ReservationIn, db: Session):
     ticket = db.query(models.Ticket).filter(models.Ticket.id == reservation.ticket_id).first()
-    if ticket.num_available >= reservation.num_tickets:
+    event = db.query(models.Event).filter(models.Event.id == ticket.event_id).first()
+    if ticket.num_available >= reservation.num_tickets and event.date_time > datetime.now():
         new_reservation = models.Reservation(ticket_id=reservation.ticket_id, num_reserved=reservation.num_tickets)
         db.add(new_reservation)
         ticket.num_available -= reservation.num_tickets
@@ -31,7 +33,9 @@ def make_reservation(reservation: schemas.ReservationIn, db: Session):
     
 def update_reservation(reservation_id: int, num_tickets: int, db: Session):
     reservation = db.query(models.Reservation).filter(models.Reservation.id == reservation_id).first()
-    if reservation:
+    ticket = db.query(models.Ticket).filter(models.Ticket.id == reservation.ticket_id).first()
+    event = db.query(models.Event).filter(models.Event.id == ticket.event_id).first()
+    if reservation and event.date_time > datetime.now():
         ticket = reservation.ticket
         num_available = ticket.num_available + reservation.num_reserved
         if num_tickets <= num_available:
@@ -47,7 +51,9 @@ def update_reservation(reservation_id: int, num_tickets: int, db: Session):
     
 def cancel_reservation(reservation_id: int, db: Session):
     reservation = db.query(models.Reservation).filter(models.Reservation.id == reservation_id).first()
-    if reservation:
+    ticket = db.query(models.Ticket).filter(models.Ticket.id == reservation.ticket_id).first()
+    event = db.query(models.Event).filter(models.Event.id == ticket.event_id).first()
+    if reservation and event.date_time > datetime.now():
         ticket = reservation.ticket
         ticket.num_available += reservation.num_reserved
         db.delete(reservation)
